+++ PAG-prueba-01/objloader.hpp
@@ -1,6 +1,7 @@
 #ifndef OBJLOADER_H
 #define OBJLOADER_H
 
+#include <GL\glew.h>
 bool loadOBJ(
 	const char * path, 
 	std::vector<glm::vec3> & out_vertices, 
@@ -12,7 +13,7 @@ bool loadOBJ(
 
 bool loadAssImp(
 	const char * path, 
-	std::vector<unsigned short> & indices,
+	std::vector<GLuint> & indices,
 	std::vector<glm::vec3> & vertices,
 	std::vector<glm::vec2> & uvs,
 	std::vector<glm::vec3> & normals
